+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Yii Framework 2 gii extension Change Log
 -----------------------
 
 - Bug #531: Fix `yii\gii\console\GenerateAction` to use `stdout()` instead of echoing the output (egmsystems)
+- Bug #532: Return `ExitCode::USAGE` on command input validation error (egmsystems)
+- Enh #537: Generating rules for the fields with default values (manky)
+- Enh #542: Use the table name to create the relation (thiagotalma)
 
 
 2.2.6 May 22, 2023